+++ etc/build_dir.sh
@@ -82,13 +82,6 @@ if [[ $KERNEL == *Msys* ]]; then
 fi
 
 if [[ $RUNME == 1 && $0 != *runme.sh* && $0 != runme_minimal.sh ]]; then
-
-    if [[ $KERNEL == *Linux* ]]; then
-	if [[ ! -e $BUILD_DIR/install/bin/cmake ]]; then
-	    ./etc/install_cmake.sh
-	fi
-    fi
-    
     echo "Build directory is ${BUILD_DIR}"
     echo "Version to build is v${mrv2_VERSION}"
     echo "Architecture is ${ARCH}"